add python3.6 support (#489)
* add python3.6 support

* fix pybboxes version for py36 support

* drop py36 tests in torch1.11

* dont install layer for py3.6 tests

* add transformers setup into ci

* better dependency managment for layer

* update transformers installation for py36 in ci

* fix layer tests

* drop huggingface support for py36

* revert changes in readme

* add python3.6 support

* drop py36 tests in torch1.11

* update torchvision version in macos ci

* fix typo in ci

* fix mmdet installation in readme

* fix huggingface  tests

* fix torch version for macos and py36

* fix torch version for py36+macos
