+++ .github/workflows/ci.yml
@@ -85,6 +85,10 @@ jobs:
         run: >
           pip install yolov5==6.1.3 norfair==1.0.0
 
+      - name: Install Transformers(4.20.0)
+        run: >
+          pip install transformers==4.20.0
+
       - name: Install Layer (0.10.*)
         run: >
           pip install -U "layer>=0.10.0,<0.11.0"
