+++ .github/workflows/ci_torch1.10.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         operating-system: [ubuntu-latest, windows-latest, macos-latest]
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.6, 3.7, 3.8, 3.9]
       fail-fast: false
 
     steps:
@@ -73,9 +73,9 @@ jobs:
           pip install torch==1.10.2+cpu torchvision==0.11.3+cpu
           -f https://download.pytorch.org/whl/torch_stable.html
 
-      - name: Install PyTorch on MacOS
+      - name: Install PyTorch(1.10.1) and TorchVision(0.11.2) on MacOS
         if: matrix.operating-system == 'macos-latest'
-        run: pip install torch==1.10.2 torchvision==0.11.3
+        run: pip install torch==1.10.1 torchvision==0.11.2
 
       - name: Install MMDetection(2.25.0) with MMCV(1.5.3)
         run: >
@@ -90,7 +90,15 @@ jobs:
         run: >
           python -m pip install detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/torch1.10/index.html
 
+      - name: Install Transformers(4.20.0)
+        if: >
+          matrix.python-version != '3.6'
+        run: >
+          pip install transformers==4.20.0
+
       - name: Install Layer (0.10.*)
+        if: >
+          matrix.python-version != '3.6'
         run: >
           pip install -U "layer>=0.10.0,<0.11.0"
