+++ tests/test_predict.py
@@ -165,93 +165,6 @@ class TestPredict(unittest.TestCase):
                 num_car += 1
         self.assertEqual(num_car, 2)
 
-    def test_get_prediction_huggingface(self):
-        from sahi.model import HuggingfaceDetectionModel
-        from sahi.predict import get_prediction
-        from sahi.utils.huggingface import HuggingfaceTestConstants
-
-        huggingface_detection_model = HuggingfaceDetectionModel(
-            model_path=HuggingfaceTestConstants.YOLOS_TINY_MODEL_PATH,
-            confidence_threshold=CONFIDENCE_THRESHOLD,
-            device=MODEL_DEVICE,
-            category_remapping=None,
-            load_at_init=False,
-            image_size=IMAGE_SIZE,
-        )
-        huggingface_detection_model.load_model()
-
-        # prepare image
-        image_path = "tests/data/small-vehicles1.jpeg"
-        image = read_image(image_path)
-
-        # get full sized prediction
-        prediction_result = get_prediction(
-            image=image,
-            detection_model=huggingface_detection_model,
-            shift_amount=[0, 0],
-            full_shape=None,
-            postprocess=None,
-        )
-        object_prediction_list = prediction_result.object_prediction_list
-
-        # compare
-        self.assertEqual(len(object_prediction_list), 28)
-        num_person = num_truck = num_car = 0
-        for object_prediction in object_prediction_list:
-            if object_prediction.category.name == "person":
-                num_person += 1
-            elif object_prediction.category.name == "truck":
-                num_truck += 1
-            elif object_prediction.category.name == "car":
-                num_car += 1
-        self.assertEqual(num_person, 0)
-        self.assertEqual(num_truck, 1)
-        self.assertEqual(num_car, 27)
-
-    def test_get_prediction_automodel_huggingface(self):
-        from sahi.auto_model import AutoDetectionModel
-        from sahi.predict import get_prediction
-        from sahi.utils.huggingface import HuggingfaceTestConstants
-
-        huggingface_detection_model = AutoDetectionModel.from_local(
-            model_type="huggingface",
-            model_path=HuggingfaceTestConstants.YOLOS_TINY_MODEL_PATH,
-            confidence_threshold=CONFIDENCE_THRESHOLD,
-            device=MODEL_DEVICE,
-            category_remapping=None,
-            load_at_init=False,
-            image_size=IMAGE_SIZE,
-        )
-        huggingface_detection_model.load_model()
-
-        # prepare image
-        image_path = "tests/data/small-vehicles1.jpeg"
-        image = read_image(image_path)
-
-        # get full sized prediction
-        prediction_result = get_prediction(
-            image=image,
-            detection_model=huggingface_detection_model,
-            shift_amount=[0, 0],
-            full_shape=None,
-            postprocess=None,
-        )
-        object_prediction_list = prediction_result.object_prediction_list
-
-        # compare
-        self.assertEqual(len(object_prediction_list), 28)
-        num_person = num_truck = num_car = 0
-        for object_prediction in object_prediction_list:
-            if object_prediction.category.name == "person":
-                num_person += 1
-            elif object_prediction.category.name == "truck":
-                num_truck += 1
-            elif object_prediction.category.name == "car":
-                num_car += 1
-        self.assertEqual(num_person, 0)
-        self.assertEqual(num_truck, 1)
-        self.assertEqual(num_car, 27)
-
     def test_get_sliced_prediction_mmdet(self):
         from sahi.model import MmdetDetectionModel
         from sahi.predict import get_sliced_prediction
@@ -380,63 +293,6 @@ class TestPredict(unittest.TestCase):
             if object_prediction.category.name == "car":
                 num_car += 1
         self.assertEqual(num_car, 10)
-
-    def test_get_sliced_prediction_huggingface(self):
-        from sahi.model import HuggingfaceDetectionModel
-        from sahi.predict import get_sliced_prediction
-        from sahi.utils.huggingface import HuggingfaceTestConstants
-
-        huggingface_detection_model = HuggingfaceDetectionModel(
-            model_path=HuggingfaceTestConstants.YOLOS_TINY_MODEL_PATH,
-            confidence_threshold=CONFIDENCE_THRESHOLD,
-            device=MODEL_DEVICE,
-            category_remapping=None,
-            load_at_init=False,
-            image_size=IMAGE_SIZE,
-        )
-        huggingface_detection_model.load_model()
-
-        # prepare image
-        image_path = "tests/data/small-vehicles1.jpeg"
-
-        slice_height = 512
-        slice_width = 512
-        overlap_height_ratio = 0.1
-        overlap_width_ratio = 0.2
-        postprocess_type = "GREEDYNMM"
-        match_metric = "IOS"
-        match_threshold = 0.5
-        class_agnostic = True
-
-        # get sliced prediction
-        prediction_result = get_sliced_prediction(
-            image=image_path,
-            detection_model=huggingface_detection_model,
-            slice_height=slice_height,
-            slice_width=slice_width,
-            overlap_height_ratio=overlap_height_ratio,
-            overlap_width_ratio=overlap_width_ratio,
-            perform_standard_pred=False,
-            postprocess_type=postprocess_type,
-            postprocess_match_threshold=match_threshold,
-            postprocess_match_metric=match_metric,
-            postprocess_class_agnostic=class_agnostic,
-        )
-        object_prediction_list = prediction_result.object_prediction_list
-
-        # compare
-        self.assertEqual(len(object_prediction_list), 53)
-        num_person = num_truck = num_car = 0
-        for object_prediction in object_prediction_list:
-            if object_prediction.category.name == "person":
-                num_person += 1
-            elif object_prediction.category.name == "truck":
-                num_truck += 1
-            elif object_prediction.category.name == "car":
-                num_car += 1
-        self.assertEqual(num_person, 0)
-        self.assertEqual(num_truck, 5)
-        self.assertEqual(num_car, 48)
 
     def test_coco_json_prediction(self):
         from sahi.predict import predict
